lib 以下を skype/bot に移動して bin を gerrit に依存しない作りにした
