+++ bin/skype-bot.rb
@@ -1,0 +1,10 @@
+#!/usr/bin/env ruby
+#coding:utf-8
+
+require File.expand_path(File.dirname(__FILE__)) + '/../config/boot.rb'
+require 'config'
+require 'skype/bot'
+
+include Skype::Bot
+
+Bots::GerritBot.new.listen_stream(Boot.config)
