--- lib/gerrit.rb
@@ -1,5 +1,0 @@
-module Gerrit
-  autoload :Streamer, 'gerrit/streamer'
-  autoload :Event,    'gerrit/event'
-end
-
