+++ lib/skype/bot.rb
@@ -1,0 +1,7 @@
+module Skype
+  module Bot
+    autoload :Bots,     'skype/bot/bots'
+    autoload :Gerrit,   'skype/bot/gerrit'
+    autoload :SkypeBot, 'skype/bot/skype_bot'
+  end
+end
