+++ lib/skype/bot/bots.rb
@@ -1,0 +1,7 @@
+module Skype
+  module Bot
+    module Bots
+      autoload :GerritBot, 'skype/bot/bots/gerrit_bot'
+    end
+  end
+end
