+++ lib/skype/bot/gerrit.rb
@@ -1,0 +1,9 @@
+module Skype
+  module Bot
+    module Gerrit
+      autoload :Streamer, 'skype/bot/gerrit/streamer'
+      autoload :Event,    'skype/bot/gerrit/event'
+    end
+  end
+end
+
